+++ src/textual/command.py
@@ -21,7 +21,16 @@ from dataclasses import dataclass
 from functools import total_ordering
 from inspect import isclass
 from time import monotonic
-from typing import TYPE_CHECKING, Any, AsyncGenerator, AsyncIterator, ClassVar, Iterable
+from typing import (
+    TYPE_CHECKING,
+    Any,
+    AsyncGenerator,
+    AsyncIterator,
+    Callable,
+    ClassVar,
+    Iterable,
+    NamedTuple,
+)
 
 import rich.repr
 from rich.align import Align
@@ -162,6 +171,9 @@ class DiscoveryHit:
 
 Hits: TypeAlias = AsyncIterator["DiscoveryHit | Hit"]
 """Return type for the command provider's `search` method."""
+
+ProviderSource: TypeAlias = "Iterable[type[Provider] | Callable[[], type[Provider]]]"
+"""The type used to declare the providers for a CommandPalette."""
 
 
 class Provider(ABC):
@@ -307,6 +319,78 @@ class Provider(ABC):
         """
 
 
+class SimpleCommand(NamedTuple):
+    """A simple command."""
+
+    name: str
+    """The name of the command."""
+    callback: IgnoreReturnCallbackType
+    """The callback to invoke when the command is selected."""
+    help_text: str | None = None
+    """The description of the command."""
+
+
+CommandListItem: TypeAlias = (
+    "SimpleCommand | tuple[str, IgnoreReturnCallbackType, str | None] | tuple[str, IgnoreReturnCallbackType]"
+)
+
+
+class SimpleProvider(Provider):
+    """A simple provider which the caller can pass commands to."""
+
+    def __init__(
+        self,
+        screen: Screen[Any],
+        commands: list[CommandListItem],
+    ) -> None:
+        # Convert all commands to SimpleCommand instances
+        super().__init__(screen, None)
+        self._commands: list[SimpleCommand] = []
+        for command in commands:
+            if isinstance(command, SimpleCommand):
+                self._commands.append(command)
+            elif len(command) == 2:
+                self._commands.append(SimpleCommand(*command, None))
+            elif len(command) == 3:
+                self._commands.append(SimpleCommand(*command))
+            else:
+                raise ValueError(f"Invalid command: {command}")
+
+    def __call__(
+        self, screen: Screen[Any], match_style: Style | None = None
+    ) -> SimpleProvider:
+        self.__match_style = match_style
+        return self
+
+    @property
+    def match_style(self) -> Style | None:
+        return self.__match_style
+
+    async def search(self, query: str) -> Hits:
+        matcher = self.matcher(query)
+        for name, callback, help_text in self._commands:
+            if (match := matcher.match(name)) > 0:
+                yield Hit(
+                    match,
+                    matcher.highlight(name),
+                    callback,
+                    help=help_text,
+                )
+
+    async def discover(self) -> Hits:
+        """Handle a request for the discovery commands for this provider.
+
+        Yields:
+            Commands that can be discovered.
+        """
+        for name, callback, help_text in self._commands:
+            yield DiscoveryHit(
+                name,
+                callback,
+                help=help_text,
+            )
+
+
 @rich.repr.auto
 @total_ordering
 class Command(Option):
@@ -349,14 +433,13 @@ class CommandList(OptionList, can_focus=False):
     CommandList {
         visibility: hidden;
         border-top: blank;
-        border-bottom: hkey $primary;
+        border-bottom: hkey $border;
         border-left: none;
         border-right: none;
         height: auto;
         max-height: 70vh;
         background: transparent;
         padding: 0;
-        text-style: bold;
     }
 
     CommandList:focus {
@@ -372,7 +455,9 @@ class CommandList(OptionList, can_focus=False):
     }
 
     CommandList > .option-list--option-highlighted {
-        background: $primary;
+        color: $block-cursor-foreground;
+        background: $block-cursor-background;
+        text-style: $block-cursor-text-style;
     }
 
     CommandList:nocolor > .option-list--option-highlighted {       
@@ -381,6 +466,7 @@ class CommandList(OptionList, can_focus=False):
 
     CommandList > .option-list--option {
         padding: 0 2;
+        color: $foreground;
     }
     """
 
@@ -416,13 +502,14 @@ class CommandInput(Input):
     CommandInput, CommandInput:focus {
         border: blank;
         width: 1fr;
-        background: transparent;
         padding-left: 0;
+        background: transparent;
+        background-tint: 0%;
     }
     """
 
 
-class CommandPalette(SystemModalScreen):
+class CommandPalette(SystemModalScreen[None]):
     """The Textual command palette."""
 
     AUTO_FOCUS = "CommandInput"
@@ -445,6 +532,7 @@ class CommandPalette(SystemModalScreen):
         min-height: 20;
     }
     CommandPalette {
+        color: $foreground;
         background: $background 60%;
         align-horizontal: center;        
 
@@ -464,8 +552,9 @@ class CommandPalette(SystemModalScreen):
     }
 
     CommandPalette > .command-palette--help-text {           
-        text-style: not bold;        
-        color: $text-muted;       
+        color: auto 50%;
+        background: transparent;
+        text-style: not bold;       
     }
 
     CommandPalette:dark > .command-palette--highlight {
@@ -486,13 +575,13 @@ class CommandPalette(SystemModalScreen):
         margin-top: 3; 
         height: 100%;
         visibility: hidden;
-        background: $primary 20%;      
+        background: $panel-darken-1;
     }
 
     CommandPalette #--input {
         height: auto;
         visibility: visible;
-        border: hkey $primary;
+        border: hkey $border;
     }
 
     CommandPalette #--input.--list-visible {
@@ -517,7 +606,7 @@ class CommandPalette(SystemModalScreen):
     CommandPalette LoadingIndicator {
         height: auto;
         visibility: hidden;
-        border-bottom: hkey $primary;
+        border-bottom: hkey $border;
     }
 
     CommandPalette LoadingIndicator.--visible {
@@ -575,9 +664,6 @@ class CommandPalette(SystemModalScreen):
     _calling_screen: var[Screen[Any] | None] = var(None)
     """A record of the screen that was active when we were called."""
 
-    _PALETTE_ID: Final[str] = "--command-palette"
-    """The internal ID for the command palette."""
-
     @dataclass
     class OptionHighlighted(Message):
         """Posted to App when an option is highlighted in the command palette."""
@@ -596,31 +682,53 @@ class CommandPalette(SystemModalScreen):
         option_selected: bool
         """True if an option was selected, False if the palette was closed without selecting an option."""
 
-    def __init__(self) -> None:
-        """Initialise the command palette."""
-        super().__init__(id=self._PALETTE_ID)
+    def __init__(
+        self,
+        providers: ProviderSource | None = None,
+        *,
+        placeholder: str = "Search for commands…",
+        name: str | None = None,
+        id: str | None = None,
+        classes: str | None = None,
+    ) -> None:
+        """Initialise the command palette.
+
+        Args:
+            providers: An optional list of providers to use. If None, the providers supplied
+                in the App or Screen will be used.
+            placeholder: The placeholder text for the command palette.
+        """
+        super().__init__(
+            id=id,
+            classes=classes,
+            name=name,
+        )
+        self.add_class("--textual-command-palette")
+
         self._selected_command: DiscoveryHit | Hit | None = None
         """The command that was selected by the user."""
         self._busy_timer: Timer | None = None
         """Keeps track of if there's a busy indication timer in effect."""
         self._no_matches_timer: Timer | None = None
         """Keeps track of if there are 'No matches found' message waiting to be displayed."""
+        self._supplied_providers: ProviderSource | None = providers
         self._providers: list[Provider] = []
         """List of Provider instances involved in searches."""
         self._hit_count: int = 0
         """Number of hits displayed."""
+        self._placeholder = placeholder
 
     @staticmethod
-    def is_open(app: App) -> bool:
-        """Is the command palette current open?
+    def is_open(app: App[object]) -> bool:
+        """Is a command palette current open?
 
         Args:
             app: The app to test.
 
         Returns:
-            `True` if the command palette is currently open, `False` if not.
+            `True` if a command palette is currently open, `False` if not.
         """
-        return app.screen.id == CommandPalette._PALETTE_ID
+        return app.screen.has_class("--textual-command-palette")
 
     @property
     def _provider_classes(self) -> set[type[Provider]]:
@@ -631,27 +739,36 @@ class CommandPalette(SystemModalScreen):
         the current screen][textual.screen.Screen.COMMANDS].
         """
 
-        def get_providers(root: App | Screen) -> Iterable[type[Provider]]:
-            """Get providers from app or screen.
+        def get_providers(
+            provider_source: ProviderSource,
+        ) -> Iterable[type[Provider]]:
+            """Load the providers from a source (typically from the COMMANDS class variable)
+            at the App or Screen level.
 
             Args:
-                root: The app or screen.
+                provider_source: The source of providers.
 
             Returns:
                 An iterable of providers.
             """
-            for provider in root.COMMANDS:
-                if isclass(provider) and issubclass(provider, Provider):
+            for provider in provider_source:
+                if isinstance(provider, SimpleProvider):
+                    yield provider
+                elif isclass(provider) and issubclass(provider, Provider):
                     yield provider
                 else:
                     # Lazy loaded providers
                     yield provider()  # type: ignore
 
-        return (
-            set()
-            if self._calling_screen is None
-            else {*get_providers(self.app), *get_providers(self._calling_screen)}
-        )
+        if self._calling_screen is None:
+            return set()
+        elif self._supplied_providers is None:
+            return {
+                *get_providers(self.app.COMMANDS),
+                *get_providers(self._calling_screen.COMMANDS),
+            }
+        else:
+            return {*get_providers(self._supplied_providers)}
 
     def compose(self) -> ComposeResult:
         """Compose the command palette.
@@ -662,7 +779,7 @@ class CommandPalette(SystemModalScreen):
         with Vertical(id="--container"):
             with Horizontal(id="--input"):
                 yield SearchIcon()
-                yield CommandInput(placeholder="Search for commands…")
+                yield CommandInput(placeholder=self._placeholder)
                 if not self.run_on_select:
                     yield Button("\u25b6")
             with Vertical(id="--results"):
@@ -689,9 +806,7 @@ class CommandPalette(SystemModalScreen):
         self.app.post_message(CommandPalette.Opened())
         self._calling_screen = self.app.screen_stack[-2]
 
-        match_style = self.get_component_rich_style(
-            "command-palette--highlight", partial=True
-        )
+        match_style = self.get_component_rich_style("command-palette--highlight")
 
         assert self._calling_screen is not None
         self._providers = [
@@ -941,7 +1056,6 @@ class CommandPalette(SystemModalScreen):
         Args:
             search_value: The value to search for.
         """
-
         # The list to hold on to the commands we've gathered from the
         # command providers.
         gathered_commands: list[Command] = []
@@ -1011,6 +1125,7 @@ class CommandPalette(SystemModalScreen):
                     yield Content.styled(hit.help, help_style)
 
             prompt = Content("\n").join(build_prompt())
+
             gathered_commands.append(Command(prompt, hit, id=str(command_id)))
 
             if worker.is_cancelled: