+++ src/textual/widgets/_sparkline.py
@@ -41,10 +41,10 @@ class Sparkline(Widget):
         height: 1;
     }
     Sparkline > .sparkline--max-color {
-        color: $accent;
+        color: $primary;
     }
     Sparkline > .sparkline--min-color {
-        color: $accent 30%;
+        color: $primary 30%;
     }
     """