+++ src/textual/widgets/_tabbed_content.py
@@ -169,7 +169,6 @@ class TabPane(Widget):
     DEFAULT_CSS = """
     TabPane {
         height: auto;
-        padding: 1 2;
     }
     """